+++ hooks/lando-setup-install-ca-wsl.js
@@ -118,17 +118,6 @@ module.exports = async (lando, options) => {
       }
     },
     canRun: async () => {
-      // get windows user
-      const user = await getWinEnvar('USERNAME');
-
-      // Check for admin privileges
-      if (!await require('../utils/is-admin-user')(user, {platform: 'win32'})) {
-        throw new Error([
-          `User "${user}" does not have permission to trust the CA!`,
-          'Contact your system admin for permission and then rerun setup.',
-        ].join(os.EOL));
-      }
-
       return true;
     },
     task: async (ctx, task) => {